requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Bcrypt==1.0.1
requests==2.28.2
python-dotenv==1.0.0
yfinance==0.2.12
pytest==7.3.1
gunicorn==20.1.0
email-validator==1.3.1
Werkzeug==2.2.3
SQLAlchemy==2.0.4 